renamed trial to cloud_trial
